# Reduce the number of syscalls in FileCache::loadMetadata